+++ src/tcl/tclutil.c
@@ -175,10 +175,9 @@ TclFindElement(interp, list, elementPtr, nextPtr, sizePtr, bracePtr)
 			/* null body */
 		    }
 		    Tcl_ResetResult(interp);
-		    /* sde cast to (int) of precision specifier */
 		    sprintf(interp->result,
 			    "list element in braces followed by \"%.*s\" instead of space",
-			    (int) (p2-p), p);
+			    p2-p, p);
 		    return TCL_ERROR;
 		} else if (openBraces != 0) {
 		    openBraces--;
@@ -233,10 +232,9 @@ TclFindElement(interp, list, elementPtr, nextPtr, sizePtr, bracePtr)
 			/* null body */
 		    }
 		    Tcl_ResetResult(interp);
-		    /* sde cast to (int) of precision specifier */
 		    sprintf(interp->result,
 			    "list element in quotes followed by \"%.*s\" %s",
-			    (int) (p2-p), p, "instead of space");
+			    p2-p, p, "instead of space");
 		    return TCL_ERROR;
 		}
 		break;
